Revert eager detection
