+++ Rakefile
@@ -3,7 +3,7 @@ require "rake/testtask"
 task :default => :test
 
 $:.unshift File.expand_path("../lib", __FILE__)
-require "execjs/runtimes"
+require "execjs"
 
 tests = namespace :test do |tests|
   ExecJS::Runtimes.names.each do |name|
