+++ test/execjs_test.rb
@@ -1,8 +1,9 @@
 require "test/unit"
-require "execjs/module"
+require "execjs"
 
 begin
   require "execjs"
+  ExecJS.runtime
 rescue ExecJS::RuntimeUnavailable => e
   warn e
   exit 2
